pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "wszebishop"
version = "0.1.0"
description = "Website where you can sell and buy products"
authors = ["wszeborj <wszeborj@o2.pl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
awsebcli = "3.21.0"
crispy-bootstrap4 = "2022.1"
Django = "4.2.4"
django-bootstrap-datepicker-plus = "5.0.4"
django-phonenumber-field = "^7.1.0"
django-plotly-dash = "2.3.1"
django-storages = "1.14.4"
factory-boy = "3.3.0"
Faker = "22.6.0"
icecream = "2.1.3"
numpy = "1.26.4"
pandas = "2.2.2"
pillow = "10.0.0"
plotly = "5.22.0"
poetry = "^1.8.4"
psycopg2-binary = "2.9.10"
requests = "2.31.0"
sentry-sdk = "2.6.0"
django-environ = "^0.11.2"
django-extensions = "^3.2.3"
django-filter = "^24.2"
phonenumberslite = "^8.13.19"
stripe = "^8.5.0"
python-box = "^7.2.0"
boto3 = "^1.35.66"

[tool.poetry.dev-dependencies]
pre-commit = "3.6.2"
black = "23.3.0"
conventional-pre-commit = "2.3.0"
isort = "5.12.0"
bandit = "1.7.5"
